requirements.txt
# Azure AI Foundry & OpenAI
openai>=1.55.0
azure-ai-projects>=1.0.0b5
azure-ai-inference>=1.0.0b6
azure-identity>=1.19.0
azure-storage-blob>=12.20.0
azure-servicebus>=7.12.0
azure-ai-vision-imageanalysis>=1.0.0

# Semantic Kernel for orchestration
semantic-kernel>=1.15.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=23.0.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
pillow>=10.4.0

# Database
azure-cosmos>=4.7.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
aiohttp>=3.10.0
requests>=2.32.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Monitoring
opencensus-ext-azure>=1.1.0
applicationinsights>=0.11.10
